Improve logging for the extract.xml action and command
Previously, the action logged a message saying that it was
extracting the xml after the computation had finished, and part
of the log message was in the action while the other part was
logged in the command. It's simpler and less confusing to log
"starting extraction" and "finished extracting" messages in the
right order and in one place.