Fix bug in amti.actions.review.review.review_hit
Simplify the code and fix a bug in amti.actions.review.review_hit
by leveraging click's built-in user input collecting features
rather than manually collecting and validating the user input. The
bug previously was that after a user marked an assignment and
specified what action they want to take on it, they'd always be
asked again what action they wanted to take (and given the option
to mark it again).